fix: use the test class/method to compute the test namespace

+++ operator-framework-junit5/src/main/java/io/javaoperatorsdk/operator/junit/OperatorExtension.java
@@ -1,30 +1,15 @@
 package io.javaoperatorsdk.operator.junit;
 
+import static io.javaoperatorsdk.operator.api.config.ControllerConfigurationOverrider.override;
+
 import io.fabric8.kubernetes.api.model.HasMetadata;
-import java.io.IOException;
-import java.io.InputStream;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.UUID;
-import java.util.concurrent.TimeUnit;
-import java.util.stream.Collectors;
-
-import org.awaitility.Awaitility;
-import org.junit.jupiter.api.extension.AfterAllCallback;
-import org.junit.jupiter.api.extension.AfterEachCallback;
-import org.junit.jupiter.api.extension.BeforeAllCallback;
-import org.junit.jupiter.api.extension.BeforeEachCallback;
-import org.junit.jupiter.api.extension.ExtensionContext;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 import io.fabric8.kubernetes.api.model.KubernetesResourceList;
 import io.fabric8.kubernetes.api.model.NamespaceBuilder;
-import io.fabric8.kubernetes.client.CustomResource;
 import io.fabric8.kubernetes.client.DefaultKubernetesClient;
 import io.fabric8.kubernetes.client.KubernetesClient;
 import io.fabric8.kubernetes.client.dsl.NonNamespaceOperation;
 import io.fabric8.kubernetes.client.dsl.Resource;
+import io.fabric8.kubernetes.client.utils.KubernetesResourceUtil;
 import io.fabric8.kubernetes.client.utils.Utils;
 import io.javaoperatorsdk.operator.Operator;
 import io.javaoperatorsdk.operator.api.ResourceController;
@@ -33,8 +18,20 @@ import io.javaoperatorsdk.operator.api.config.ConfigurationService;
 import io.javaoperatorsdk.operator.api.config.Version;
 import io.javaoperatorsdk.operator.processing.ConfiguredController;
 import io.javaoperatorsdk.operator.processing.retry.Retry;
-
-import static io.javaoperatorsdk.operator.api.config.ControllerConfigurationOverrider.override;
+import java.io.InputStream;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Locale;
+import java.util.concurrent.TimeUnit;
+import java.util.stream.Collectors;
+import org.awaitility.Awaitility;
+import org.junit.jupiter.api.extension.AfterAllCallback;
+import org.junit.jupiter.api.extension.AfterEachCallback;
+import org.junit.jupiter.api.extension.BeforeAllCallback;
+import org.junit.jupiter.api.extension.BeforeEachCallback;
+import org.junit.jupiter.api.extension.ExtensionContext;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 public class OperatorExtension
     implements HasKubernetesClient,
@@ -47,17 +44,22 @@ public class OperatorExtension
 
   private final KubernetesClient kubernetesClient;
   private final ConfigurationService configurationService;
-  private final String namespace;
   private final Operator operator;
+  private final List<ControllerSpec> controllers;
   private final boolean preserveNamespaceOnError;
   private final boolean waitForNamespaceDeletion;
 
+  private String namespace;
+
   private OperatorExtension(
-      ConfigurationService configurationService, boolean preserveNamespaceOnError,
+      ConfigurationService configurationService,
+      List<ControllerSpec> controllers,
+      boolean preserveNamespaceOnError,
       boolean waitForNamespaceDeletion) {
+
     this.kubernetesClient = new DefaultKubernetesClient();
-    this.namespace = UUID.randomUUID().toString();
     this.configurationService = configurationService;
+    this.controllers = controllers;
     this.operator = new Operator(this.kubernetesClient, this.configurationService);
     this.preserveNamespaceOnError = preserveNamespaceOnError;
     this.waitForNamespaceDeletion = waitForNamespaceDeletion;
@@ -127,57 +129,60 @@ public class OperatorExtension
     return kubernetesClient.resources(type).inNamespace(namespace).create(resource);
   }
 
-  public void register(ResourceController<? extends CustomResource<?, ?>> controller) {
-    register(controller, null);
-  }
-
-  @SuppressWarnings({"unchecked", "rawtypes"})
-  public void register(ResourceController controller, Retry retry) {
-    final var config = configurationService.getConfigurationFor(controller);
-    final var oconfig = override(config).settingNamespace(namespace);
-    final var path = "/META-INF/fabric8/" + config.getCRDName() + "-v1.yml";
-
-    if (retry != null) {
-      oconfig.withRetry(retry);
-    }
-
-    try (InputStream is = getClass().getResourceAsStream(path)) {
-      kubernetesClient.load(is).createOrReplace();
-    } catch (Exception ex) {
-      throw new IllegalStateException("Cannot apply CRD yaml: " + path, ex);
-    }
-
-    if (controller instanceof KubernetesClientAware) {
-      ((KubernetesClientAware) controller).setKubernetesClient(kubernetesClient);
-    }
-
-    this.operator.register(controller, oconfig.build());
-
-    LOGGER.info("Controller {} is registered", controller.getClass().getCanonicalName());
-  }
-
+  @SuppressWarnings("unchecked")
   protected void before(ExtensionContext context) {
+    namespace = context.getRequiredTestClass().getSimpleName();
+    namespace += "-";
+    namespace += context.getRequiredTestMethod().getName();
+    namespace = KubernetesResourceUtil.sanitizeName(namespace).toLowerCase(Locale.US);
+
     LOGGER.info("Initializing integration test in namespace {}", namespace);
 
     kubernetesClient
         .namespaces()
         .create(new NamespaceBuilder().withNewMetadata().withName(namespace).endMetadata().build());
 
+
+    for (var ref : controllers) {
+      final var config = configurationService.getConfigurationFor(ref.controller);
+      final var oconfig = override(config).settingNamespace(namespace);
+      final var path = "/META-INF/fabric8/" + config.getCRDName() + "-v1.yml";
+
+      if (ref.retry != null) {
+        oconfig.withRetry(ref.retry);
+      }
+
+      try (InputStream is = getClass().getResourceAsStream(path)) {
+        kubernetesClient.load(is).createOrReplace();
+      } catch (Exception ex) {
+        throw new IllegalStateException("Cannot apply CRD yaml: " + path, ex);
+      }
+
+      if (ref.controller instanceof KubernetesClientAware) {
+        ((KubernetesClientAware) ref.controller).setKubernetesClient(kubernetesClient);
+      }
+
+
+      this.operator.register(ref.controller, oconfig.build());
+    }
+
     this.operator.start();
   }
 
   protected void after(ExtensionContext context) {
-    if (preserveNamespaceOnError && context.getExecutionException().isPresent()) {
-      LOGGER.info("Preserving namespace {}", namespace);
-    } else {
-      LOGGER.info("Deleting namespace {} and stopping operator", namespace);
-      kubernetesClient.namespaces().withName(namespace).delete();
-      if (waitForNamespaceDeletion) {
-        LOGGER.info("Waiting for namespace {} to be deleted", namespace);
-        Awaitility.await("namespace deleted")
-            .pollInterval(50, TimeUnit.MILLISECONDS)
-            .atMost(60, TimeUnit.SECONDS)
-            .until(() -> kubernetesClient.namespaces().withName(namespace).get() == null);
+    if (namespace != null) {
+      if (preserveNamespaceOnError && context.getExecutionException().isPresent()) {
+        LOGGER.info("Preserving namespace {}", namespace);
+      } else {
+        LOGGER.info("Deleting namespace {} and stopping operator", namespace);
+        kubernetesClient.namespaces().withName(namespace).delete();
+        if (waitForNamespaceDeletion) {
+          LOGGER.info("Waiting for namespace {} to be deleted", namespace);
+          Awaitility.await("namespace deleted")
+              .pollInterval(50, TimeUnit.MILLISECONDS)
+              .atMost(60, TimeUnit.SECONDS)
+              .until(() -> kubernetesClient.namespaces().withName(namespace).get() == null);
+        }
       }
     }
 
@@ -245,28 +250,25 @@ public class OperatorExtension
       return this;
     }
 
-    @SuppressWarnings({"rawtypes", "unchecked"})
     public OperatorExtension build() {
-      OperatorExtension answer =
-          new OperatorExtension(configurationService, preserveNamespaceOnError,
-              waitForNamespaceDeletion);
-      for (ControllerSpec spec : controllers) {
-        answer.register(spec.controller, spec.retry);
-      }
-
-      return answer;
+      return new OperatorExtension(
+          configurationService,
+          controllers,
+          preserveNamespaceOnError,
+          waitForNamespaceDeletion);
     }
+  }
 
-    private static class ControllerSpec {
-      final ResourceController controller;
-      final Retry retry;
+  @SuppressWarnings("rawtypes")
+  private static class ControllerSpec {
+    final ResourceController controller;
+    final Retry retry;
 
-      public ControllerSpec(
-          ResourceController controller,
-          Retry retry) {
-        this.controller = controller;
-        this.retry = retry;
-      }
+    public ControllerSpec(
+        ResourceController controller,
+        Retry retry) {
+      this.controller = controller;
+      this.retry = retry;
     }
   }
 }
